honor config in stage_add_all aswell
