+++ asyncgit/src/sync/utils.rs
@@ -1,7 +1,10 @@
 //! sync git api (various methods)
 
 use super::CommitId;
-use crate::error::{Error, Result};
+use crate::{
+    error::{Error, Result},
+    sync::config::untracked_files_config,
+};
 use git2::{IndexAddOption, Repository, RepositoryOpenFlags};
 use scopetime::scope_time;
 use std::{
@@ -129,7 +132,18 @@ pub fn stage_add_all(repo_path: &str, pattern: &str) -> Result<()> {
 
     let mut index = repo.index()?;
 
-    index.add_all(vec![pattern], IndexAddOption::DEFAULT, None)?;
+    let config = untracked_files_config(&repo)?;
+
+    if config.include_untracked() || config.recurse_untracked_dirs() {
+        index.add_all(
+            vec![pattern],
+            IndexAddOption::DEFAULT,
+            None,
+        )?;
+    } else {
+        index.update_all(vec![pattern], None)?;
+    }
+
     index.write()?;
 
     Ok(())
@@ -147,37 +161,6 @@ pub fn stage_addremoved(repo_path: &str, path: &Path) -> Result<()> {
     index.write()?;
 
     Ok(())
-}
-
-/// get string from config
-pub fn get_config_string(
-    repo_path: &str,
-    key: &str,
-) -> Result<Option<String>> {
-    let repo = repo(repo_path)?;
-    get_config_string_repo(&repo, key)
-}
-
-pub(crate) fn get_config_string_repo(
-    repo: &Repository,
-    key: &str,
-) -> Result<Option<String>> {
-    let cfg = repo.config()?;
-
-    // this code doesnt match what the doc says regarding what
-    // gets returned when but it actually works
-    let entry_res = cfg.get_entry(key);
-
-    let entry = match entry_res {
-        Ok(ent) => ent,
-        Err(_) => return Ok(None),
-    };
-
-    if entry.has_value() {
-        Ok(entry.value().map(std::string::ToString::to_string))
-    } else {
-        Ok(None)
-    }
 }
 
 pub(crate) fn bytes2string(bytes: &[u8]) -> Result<String> {
@@ -246,23 +229,7 @@ mod tests {
             false
         );
     }
-    #[test]
-    fn test_get_config() {
-        let bad_dir_cfg =
-            get_config_string("oodly_noodly", "this.doesnt.exist");
-        assert!(bad_dir_cfg.is_err());
 
-        let (_td, repo) = repo_init().unwrap();
-        let path = repo.path();
-        let rpath = path.as_os_str().to_str().unwrap();
-        let bad_cfg = get_config_string(rpath, "this.doesnt.exist");
-        assert!(bad_cfg.is_ok());
-        assert!(bad_cfg.unwrap().is_none());
-        // repo init sets user.name
-        let good_cfg = get_config_string(rpath, "user.name");
-        assert!(good_cfg.is_ok());
-        assert!(good_cfg.unwrap().is_some());
-    }
     #[test]
     fn test_staging_one_file() {
         let file_path = Path::new("file1.txt");
